not all controllers may have all generic methods

+++ controller.c
@@ -100,6 +100,9 @@ get_status(worker *w, object *params)
 	// NO auth check.  Should we require authentication here?
 
 	if (get_controller_param(w, params, &cp)) {
+		if (!cp->ops->get_status) {
+			return;
+		}
 		cp->ops->get_status(cp, w);
 	}
 }
@@ -111,6 +114,9 @@ get_networks(worker *w, object *params)
 
 	if (get_auth_param(w, params, NULL) &&
 	    get_controller_param(w, params, &cp)) {
+		if (!cp->ops->get_networks) {
+			return;
+		}
 		cp->ops->get_networks(cp, w);
 	}
 }
@@ -123,6 +129,9 @@ get_network(worker *w, object *params)
 
 	if (get_auth_param(w, params, NULL) &&
 	    get_network_param(w, params, &cp, &nwid)) {
+		if (!cp->ops->get_network) {
+			return;
+		}
 		cp->ops->get_network(cp, w, nwid);
 	}
 }
@@ -135,6 +144,9 @@ get_network_members(worker *w, object *params)
 
 	if (get_auth_param(w, params, NULL) &&
 	    get_network_param(w, params, &cp, &nwid)) {
+		if (!cp->ops->get_members) {
+			return;
+		}
 		cp->ops->get_members(cp, w, nwid);
 	}
 }
@@ -148,6 +160,9 @@ get_network_member(worker *w, object *params)
 
 	if (get_auth_param(w, params, NULL) &&
 	    get_member_param(w, params, &cp, &nwid, &member)) {
+		if (!cp->ops->get_member) {
+			return;
+		}
 		cp->ops->get_member(cp, w, nwid, member);
 	}
 }
@@ -161,6 +176,9 @@ delete_network_member(worker *w, object *params)
 
 	if (get_auth_param(w, params, NULL) &&
 	    get_member_param(w, params, &cp, &nwid, &member)) {
+		if (!cp->ops->delete_member) {
+			return;
+		}
 		cp->ops->delete_member(cp, w, nwid, member);
 	}
 }
@@ -174,6 +192,9 @@ authorize_network_member(worker *w, object *params)
 
 	if (get_auth_param(w, params, NULL) &&
 	    get_member_param(w, params, &cp, &nwid, &member)) {
+		if (!cp->ops->authorize_member) {
+			return;
+		}
 		cp->ops->authorize_member(cp, w, nwid, member);
 	}
 }
@@ -187,6 +208,9 @@ deauthorize_network_member(worker *w, object *params)
 
 	if (get_auth_param(w, params, NULL) &&
 	    get_member_param(w, params, &cp, &nwid, &member)) {
+		if (!cp->ops->deauthorize_member) {
+			return;
+		}
 		cp->ops->deauthorize_member(cp, w, nwid, member);
 	}
 }
